refactor(app/renderer/reducers): move `copied-to-clipboard`
